chore(http-client): reduce the quality factor of application/json
In RFC2616 section 14.1 an accept parameter q is specified which
'indicate the relative degree of preference' of a media type in
the accept header. Since Wildfly doesn't seem to respect the ordering
we can workaround the wrong response media type selection by reduce
the quality factor of the less favored media type 'application/json'.
This should not have a negative effect on any other application
server because it is part of the HTTP spec.

related to #CAM-2718